refactor: 코드리뷰 반영
- queryRepository 사용하지 않도록 개선
- 서비스 내 repository 정렬 변경 (for 가독성)

related issue #548
