+++ src/main/java/org/sopt/makers/internal/community/repository/CommunityPostRepositoryCustom.java
@@ -1,0 +1,8 @@
+package org.sopt.makers.internal.community.repository;
+
+import org.sopt.makers.internal.dto.community.PostCategoryDao;
+
+public interface CommunityPostRepositoryCustom {
+
+	PostCategoryDao findRecentPostByCategory(String categoryName);
+}
